pyproject.toml
[tool.poetry]
name = "authenticator"
version = "0.1.0"
description = ""
authors = ["Goutham Pratapa <pratapagoutham@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
requests = "^2.25.1"
oauthlib = "^3.1.1"
Flask-Login = "^0.5.0"
click = "^8.0.1"
Flask-PyMongo = "^2.3.0"
cryptocode = "^0.1"
black = "^21.7b0"
Authlib = "^0.15.4"
google-auth = "^1.34.0"
google-auth-oauthlib = "^0.4.5"
google = "^3.0.0"
cachetools = "^4.2.2"
CacheControl = "^0.12.6"
"oslo.log" = "^4.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
